Created script to seed data base
